Working toward a new PR which includes s3 storage
