+++ Package.swift
@@ -21,14 +21,17 @@ let package = Package(
     name: "nimbella-sdk",
     platforms: [.iOS("13.0"), .macOS("10.15")],
     products: [
-        .library(name: "nimbella-sdk", targets: ["nimbella-sdk"])
+        .library(name: "nimbella-key-value", targets: ["nimbella-key-value"]),
+        .library(name: "nimbella-object", targets: ["nimbella-object"])
     ],
     dependencies: [
-        .package(url: "https://github.com/swiftpackages/DotEnv.git", from: "2.0.0"),
-        .package(url: "https://github.com/apple/swift-nio.git", from: "2.0.0")
+        .package(url: "https://github.com/apple/swift-nio.git", from: "2.0.0"),
+        .package(url: "https://gitlab.com/mordil/RediStack.git", .branch("master")),
+        .package(url: "https://github.com/swiftpackages/DotEnv.git", from: "2.0.0")
     ],
     targets: [
-        .target(name: "nimbella-sdk", dependencies: [.product(name: "NIOCore", package: "swift-nio")]),
-        .testTarget(name: "nimbella-sdk-tests", dependencies: ["nimbella-sdk", "DotEnv"])
+        .target(name: "nimbella-object", dependencies: [.product(name: "NIOCore", package: "swift-nio")]),
+        .target(name: "nimbella-key-value", dependencies: ["RediStack"]),
+        .testTarget(name: "nimbella-sdk-tests", dependencies: ["nimbella-object", "nimbella-key-value", "DotEnv"])
     ]
 )
