+++ Sources/nimbella-key-value/NimbellaRedis.swift
@@ -1,0 +1,45 @@
+/**
+ * Copyright (c) 2021-present, Joshua Auerbach
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// The visible entry point to the Nimbella SDK for redis
+
+import RediStack
+import Foundation
+import NIO
+
+
+// Errors that can occur in the Nimbella Key value SDK
+public enum NimbellaKeyValueError : Error, Equatable {
+    case notImplemented(String)
+    case noKeyValueStore
+}
+
+// Retrieve a redis client handle
+public func redis() throws -> RedisClient {
+    let env = ProcessInfo.processInfo.environment
+    guard let redisHost = env["__NIM_REDIS_IP"], !redisHost.isEmpty else {
+        throw NimbellaKeyValueError.noKeyValueStore
+    }
+    guard let redisPassword = env["__NIM_REDIS_PASSWORD"], !redisPassword.isEmpty else {
+        throw NimbellaKeyValueError.noKeyValueStore
+    }
+    let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
+    let eventLoop = eventLoopGroup.next()
+    let client = try RedisConnection.make(
+        configuration: try .init(hostname: redisHost, port: 6379, password: redisPassword),
+        boundEventLoop: eventLoop).wait()
+    return client
+}
