+++ Sources/nimbella-object/NimbellaError.swift
